third_party/WebKit/LayoutTests/images/onload-event-when-reloading-image-after-interrupted-null-src-load.html

<!DOCTYPE html>
<title>crbug.com/469678: We should fire an image load event even if the image is the last one we loaded successfully.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var img = new Image();
    var testImage = 'resources/green.jpg';
    var resetImage = '';
    // In this test we load an image. Then we make a null request (img.src='')
    // which should clear down the image before we attempt to load the same image again.
    img.onload = t.step_func(function() {
        img.src = resetImage;
        img.src = testImage;
        img.onload = t.step_func_done();
    });
    img.src = resetImage;
    img.src = testImage;
});
</script>
