third_party/WebKit/LayoutTests/platform/win/images/onload-event-when-reloading-image-after-interrupted-null-src-load-expected.txt

This is a testharness.js-based test.
PASS crbug.com/469678: We should fire an image load event even if the image is the last one we loaded successfully. 
Harness: the test ran to completion.
